various Improvements
Variables:
- added class option "playoutAnimations" -> when scrolling past the
animation should they be played out (true) or just be jumped to the
respective last frame (false)? Does not affect animations where
duration = 0
- added private var "scrollContainerOffset" -> needed for mobile
implementations as usually a scroll container is needed. Documentation
will follow.
- renamed private var "didScrollCheck" to "doUpdateOnNextTick" -> this
is merely cosmetic, as I found the previous name to suggest that
something HAS HAPPENED, when in fact something WILL HAPPEN on next tick.

Methods:
- added private function "cssNumericPosition" -> it gets position
values set in css instead of the jquery position() or offset() returns.
This is needed for Pinned elements that are moved (animated) during pin.
- added public function "setScrollContainerOffset" -> setter function
for above mentioned Variable for mobile applications. Documentation
will follow.
- changed addPin
	-> added ne option "pushFollowers" used to define wether the pinned
element should "push down" following elements or scroll past them
	-> made spacers for every pin, also changed classname of spacers to
"superscrollorama-pin-spacer" to avoid collision with existing elements
in the dom
- changed "checkScrollAnim"
	-> changed a lot of stuff regarding spacers and animation regarding
the new pushFollowers option for pins.
	-> Animations with duration 0 will now be played without regard to
scroll position, just based on the time option passed to the tween
object
	-> added suppert for indefinitely repeated animations (TweenMax option
repeat = -1). They will be repeated within the animation/pin duration
or, if duration is 0 they will play below the start point and stop
above.
	-> added onPin / onUnpin parameters. They receive a Boolean Variable,
defining wether the callback was fired at the end (true) or the
beginning (false) of the pin.