Removed function as triggerHook and changed type detection
I decided it doesn't make sense to be able to pass function as
triggerHook. Firstly it should be available to all parameters if
enabled, but shouldn't be as all parameters can be modified on the fly
from the outside.
This is far preferable for performance reasons, because a passed
function would need to be executed every time the value is needed. It's
far better to set it from the outside only when it changes.