major changes
removed addNewScene (too confusing)
and instead unified usage of controller.addScene(scene) and
Scene.addTo(controller)
Also unified all options to be settable on the fly.
corrected a small mistake when verifying triggerHook setting.
removed set parent of scene because the parent is automatically set
now, when adding the scene to the controller.