Bugfix for mobile applications
When using iScroll the calculation of the scroll position was actually
wrong, because instead of really scrolling the container, iScroll moves
the content element.
To fix this a new method was added to the controller, that allows for
custom functions for the scroll position calculation.
When adding iScroll users should now define a function to calculate the
correct scroll position.
Example:
controller.scrollPos(function () {
	return -myScroll.y;
});

I also updated the main demo page to be mobile compatible.