added: note about competing pins
an element is now defined as being able to only be pinned once at a
time.
This is the answer to the long term bug of pins not working when used
in multiple scroll directions at the same time.
This bug can not be resolved unless with big performance drawbacks.
Since this is a very rare usecase this is solved in this manner.