Fixed centered pin offset problem
If a pin is horizontally centered and its width is fixed (or auto
width), taking it out of document flow by setting its position to fixed
would cause the pin spacer to collapse to 0.

The left offset of the pin spacer moves to the right because its width
is now zero. This causes the pin to jump to the right as well.